refactor: align manager workflow with shared modules

+++ src/main/java/com/itc/employeeleaveattendance/service/LeaveRequestService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface LeaveRequestService {
 
-    List<PendingLeaveRequestDTO> getPendingRequestsForManager(int managerId);
+    List<PendingLeaveRequestDTO> getPendingRequestsForManager(long managerId);
 
-    void approveLeave(int requestId, int managerId);
+    void approveLeave(long requestId, long managerId);
 
-    void rejectLeave(int requestId, int managerId);
+    void rejectLeave(long requestId, long managerId);
 }
